Add Hermes as submodule and update CMakeLists files
This commit adds the Hermes JavaScript engine as a git submodule in the vendor directory. Changes have been made to the CMakeLists.txt files to integrate the new submodule and properly link the libraries within the build system. Additionally, a new line has been added to .gitignore to exclude build files.