Implement timer logic in JavaScript and integrate JS files into C++ code
This commit introduces new JavaScript files to implement a timer logic, replacing a previously deleted CMakeList.txt. It also integrates these JavaScript files as functions in the main C++ application through hermes JavaScript engine. Significant changes are done to CMakeList.txt in root and src directories to integrate this change and ensure the successful build of the project. Lastly, refactorings and additions to the main.cpp file were made to correspond to these changes.